Fix test and Constraints
