+++ Ted Talk/Ted TalkTests/Parse_Test.swift
@@ -1,16 +1,16 @@
 //
-//  Ted_TalkTests.swift
+//  Parse_Test.swift
 //  Ted TalkTests
 //
-//  Created by Gonzalo Perretti on 4/6/21.
+//  Created by Agustin Godoy Cosser on 02/09/2022.
 //
 
 import XCTest
 @testable import Ted_Talk
 
-class Ted_TalkTests: XCTestCase {
-    
-    // MARK: - Test init
+class Parse_Test: XCTestCase {
+
+    let sut: Parser = Parser()
     
     override func setUpWithError() throws {
         try super.setUpWithError()
@@ -20,10 +20,7 @@ class Ted_TalkTests: XCTestCase {
         try super.tearDownWithError()
     }
     
-    // MARK: - Test parse
-
     func testParse() {
-        let sut: Parser = Parser()
         let promise = self.expectation(description: "Scaling")
         var talks: [TedTalk]?
 
@@ -46,7 +43,6 @@ class Ted_TalkTests: XCTestCase {
     }
 
     func  testParseToFail() {
-        let sut: Parser = Parser()
         let promise = self.expectation(description: "Scaling")
 
         sut.parse("xxxx") { result in
@@ -60,23 +56,5 @@ class Ted_TalkTests: XCTestCase {
         
         waitForExpectations(timeout: 5, handler: nil)
     }
-    
-    // MARK: - Test searchWord
-    
-    func testSearchWord() {
-        let sut: DataManager = DataManager("test")
-        let promise = self.expectation(description: "Scaling")
-        var talks: [TedTalk]?
-        
-        sut.getTalks { result in
-            talks = result
-            promise.fulfill()
-        }
-        
-        waitForExpectations(timeout: 5, handler: nil)
-        
-        talks = sut.serchWord(searchText: "x", picker: "description")
-        
-        XCTAssertEqual(talks?.count, 0, "Score computed from guess is wrong")
-    }
+
 }
